refactor(goner): rename interfaces to g and update related packages
- Rename interfaces package to g
- Update cmux package to use new g package- Update nacos, schedule, and other packages to use new g package- Remove deprecated Priest function from cmux and schedule packages
- Add MatchFor function to cmux package
- Update go.mod and go.sum files across multiple packages
