+++ cmux/go.mod
@@ -4,11 +4,14 @@ go 1.24.1
 
 require (
 	github.com/gone-io/gone/v2 v2.0.7
+	github.com/gone-io/goner/g v0.0.0-00010101000000-000000000000
 	github.com/gone-io/goner/tracer v0.0.1
 	github.com/soheilhy/cmux v0.1.5
 	github.com/stretchr/testify v1.10.0
 	go.uber.org/mock v0.5.0
 )
+
+replace github.com/gone-io/goner/g => ../g
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
