+++ examples/gin_gorm_viper/go.mod
@@ -3,7 +3,7 @@ module gin_gorm_viper
 go 1.24.1
 
 require (
-	github.com/gone-io/gone/v2 v2.0.6
+	github.com/gone-io/gone/v2 v2.0.7
 	github.com/gone-io/goner v0.0.3
 	github.com/gone-io/goner/gin v0.0.0-20250321103707-52f1be0edf49
 	github.com/gone-io/goner/gorm v0.0.0-00010101000000-000000000000
