+++ g/go.mod
@@ -1,0 +1,3 @@
+module github.com/gone-io/goner/g
+
+go 1.24.1
