+++ g/locker.go
@@ -1,0 +1,7 @@
+package g
+
+import "time"
+
+type DoLocker interface {
+	LockAndDo(key string, fn func(), lockTime, checkPeriod time.Duration) (err error)
+}
