+++ g/router.go
@@ -1,4 +1,4 @@
-package svc
+package g
 
 type ServiceRouter interface {
 	GetServiceAddress(serviceName string) (serviceAddress string, err error)
