--- interfaces/go.mod
@@ -1,3 +1,0 @@
-module github.com/gone-io/goner/interfaces
-
-go 1.24.1
