+++ nacos/name_client.go
@@ -3,7 +3,7 @@ package nacos
 import (
 	"fmt"
 	"github.com/gone-io/gone/v2"
-	"github.com/gone-io/goner/interfaces/svc"
+	svc "github.com/gone-io/goner/g"
 	"github.com/gone-io/goner/nacos/internal/json"
 	"github.com/nacos-group/nacos-sdk-go/v2/clients"
 	"github.com/nacos-group/nacos-sdk-go/v2/clients/naming_client"
